Restructure project to allow adding integration tests
The current setup would have only allowed having all integration tests
in a single file, which we would like to avoid. Instead, we have a test
main that sets up the database, allowing for instances of the server to
be spun up by each test, which can then easily test the server.